go from f32 to f64

+++ src/camera.rs
@@ -1,4 +1,4 @@
-use std::{f32::INFINITY, io::stdout};
+use std::{f64::INFINITY, io::stdout};
 
 use rand::{thread_rng, Rng};
 
@@ -6,7 +6,7 @@ use crate::{color::{write_color, Color}, hittable::{HitRecord, Hittable}, interv
 
 #[derive(Default)]
 pub struct Camera {
-	pub aspect_ratio: f32,
+	pub aspect_ratio: f64,
 	pub image_width: i32,
 	image_height: i32,
 	center: Point3,
@@ -15,12 +15,12 @@ pub struct Camera {
 	pixel_delta_v: Vec3,
 
 	pub samples_per_pixel: i32,
-	pixel_samples_scale: f32,
+	pixel_samples_scale: f64,
 
 	pub max_depth: i32,
 }
 impl Camera {
-	pub fn new(aspect_ratio: f32, image_width: i32) -> Self {
+	pub fn new(aspect_ratio: f64, image_width: i32) -> Self {
 		Self {aspect_ratio, image_width,samples_per_pixel: 10, ..Default::default()}
 	}
 	pub fn render(&mut self, world: &dyn Hittable) {
@@ -41,21 +41,21 @@ impl Camera {
     eprintln!("\rDone.                 \n");
 	}
 	fn initialize(&mut self) {
-		self.image_height = (self.image_width as f32/self.aspect_ratio) as i32;
+		self.image_height = (self.image_width as f64/self.aspect_ratio) as i32;
 		self.image_height = if self.image_height < 1 {1} else {self.image_height};
 
-		self.pixel_samples_scale = 1.0 / self.samples_per_pixel as f32;
+		self.pixel_samples_scale = 1.0 / self.samples_per_pixel as f64;
 
 		self.center = Point3::from(0.0, 0.0, 0.0);
-		let focal_length = 1.0f32;
-		let viewport_height = 2.0f32;
-		let viewport_width = viewport_height * (self.image_width as f32/self.image_height as f32);
+		let focal_length = 1.0;
+		let viewport_height = 2.0;
+		let viewport_width = viewport_height * (self.image_width as f64/self.image_height as f64);
 
 		let viewport_u = Vec3::from(viewport_width, 0.0, 0.0);
 		let viewport_v = Vec3::from(0.0, -viewport_height, 0.0);
 
-		self.pixel_delta_u = viewport_u / self.image_width as f32;
-		self.pixel_delta_v = viewport_v / self.image_height as f32;
+		self.pixel_delta_u = viewport_u / self.image_width as f64;
+		self.pixel_delta_v = viewport_v / self.image_height as f64;
 
 		let viewport_upper_left = self.center - Vec3::from(0.0, 0.0, focal_length) - viewport_u/2.0 - viewport_v/2.0;
 		self.pixel00_loc = viewport_upper_left + 0.5 * (self.pixel_delta_u + self.pixel_delta_v);
@@ -76,8 +76,8 @@ impl Camera {
 	fn get_ray(&mut self, i: i32, j: i32) -> Ray {
 		let offset = Self::sample_square();
 		let pixel_sample = self.pixel00_loc
-			+ ((i as f32 + offset.x()) * self.pixel_delta_u)
-			+ ((j as f32 + offset.y()) * self.pixel_delta_v);
+			+ ((i as f64 + offset.x()) * self.pixel_delta_u)
+			+ ((j as f64 + offset.y()) * self.pixel_delta_v);
 		let ray_origin = self.center;
 		let ray_direction = pixel_sample - ray_origin;
 		Ray::new(ray_origin, ray_direction)
