+++ src/hittable.rs
@@ -4,7 +4,7 @@ use crate::{interval::Interval, ray::Ray, vec3::{dot, Point3, Vec3}};
 pub struct HitRecord {
 	pub p: Point3,
 	pub normal: Vec3,
-	pub t: f32,
+	pub t: f64,
 	pub front_face: bool,
 }
 impl HitRecord {
